SmartAssist/requirements.txt
# SmartAssist Python Dependencies
# Compatible with Python 3.8+ on NVIDIA Jetson Orin
# Install: pip3 install -r requirements.txt

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# Scientific computing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0

# Configuration management
PyYAML>=6.0

# Computer vision (needed for CSI operations)
opencv-python>=4.5.0

# ============================================================================
# CAN BUS COMMUNICATION
# ============================================================================

python-can>=4.0.0
cantools>=38.0.0

# ============================================================================
# CLOUD STORAGE
# ============================================================================

# Azure storage (for log upload)
azure-storage-blob>=12.14.0

# ============================================================================
# UTILITIES
# ============================================================================

python-dateutil>=2.8.0
pytz>=2021.3

# ============================================================================
# SYSTEM DEPENDENCIES (Cannot be pip installed)
# ============================================================================
# 
# GStreamer Python bindings:
#   sudo apt-get install python3-gi python3-gi-cairo
#   sudo apt-get install gir1.2-gst-1.0 gir1.2-glib-2.0
#
# GStreamer plugins:
#   sudo apt-get install gstreamer1.0-tools
#   sudo apt-get install gstreamer1.0-plugins-base
#   sudo apt-get install gstreamer1.0-plugins-good
#   sudo apt-get install gstreamer1.0-plugins-bad
#   sudo apt-get install gstreamer1.0-plugins-ugly
#   sudo apt-get install gstreamer1.0-libav
#
# GStreamer RTSP server:
#   sudo apt-get install libgstrtspserver-1.0-dev
#
# Scientific computing libraries:
#   sudo apt-get install libopenblas-base
#   sudo apt-get install libopenmpi-dev
#   sudo apt-get install libomp-dev
#
# CAN utilities:
#   sudo apt-get install can-utils
#
# ============================================================================
# NVIDIA DEPENDENCIES (Pre-installed on Jetson or via SDK Manager)
# ============================================================================
#
# - NVIDIA DeepStream SDK 6.4+
# - CUDA 12.2+
# - TensorRT 8.6+
# - PyDS (DeepStream Python bindings)
# - VPI (Vision Programming Interface)
#
# Install DeepStream:
#   Download from: https://developer.nvidia.com/deepstream-sdk
#   Or use SDK Manager: sdkmanager
#
# Verify installation:
#   deepstream-app --version
#   python3 -c "import pyds; print(pyds.__version__)"
#
# ============================================================================
# OPTIONAL: Torch2TRT (for model conversion)
# ============================================================================
#
# If you need to convert PyTorch models to TensorRT:
#   git clone https://github.com/NVIDIA-AI-IOT/torch2trt
#   cd torch2trt
#   python3 setup.py install
#
# ============================================================================

# Development dependencies (install with: pip install -r requirements.txt[dev])
# pytest>=7.0.0
# pytest-cov>=3.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950